add locking, require to be in kustomize step
